% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomWater-package.R
\docType{package}
\name{bomWater-package}
\alias{bomWater}
\alias{bomWater-package}
\title{bomWater: Download Australian Bureau of Meteorology Water Data}
\description{
An R interface to the Australian Bureau of Meteorology Water Data Online web service.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/buzacott/bomWater}
}

}
\keyword{internal}
