% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{bw_interpolation_types}
\alias{bw_interpolation_types}
\title{Access the interplation types conversion table}
\usage{
bw_interpolation_types()
}
\value{
A tibble with three columns: Interpolation Type, Name, and
Description.
}
\description{
\code{interpolation_types()} returns a table which can be used to interpret the
interpolation types returned by the WISKI API.
}
\examples{
interpolation_types()
}
