% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{bw_parameters}
\alias{bw_parameters}
\title{Available water parameters}
\source{
Interpolation types are detailed in the
BOM SOS2 manual](http://www.bom.gov.au/waterdata/wiski-web-public/Guide\\%20to\\%20Sensor\\%20Observation\\%20Services\\%20(SOS2)\\%20for\\%20Water\\%20Data\\%20\\%20Online\\%20v1.0.1.pdf)
(6.3.2, pg 40).

The conversion of SOS2 quality codes into BOM quality codes is from the
\href{http://www.bom.gov.au/waterdata/wiski-web-public/Guide\\%20to\\%20Sensor\\%20Observation\\%20Services\\%20(SOS2)\\%20for\\%20Water\\%20Data\\%20\\%20Online\\%20v1.0.1.pdf}{BOM SOS2 manual}
(6.3.3, pg 41). Quality code descriptions are from
\href{http://www.bom.gov.au/water/hrs/qc_doc.shtml}{Streamflow data quality codes for Hydrologic Reference Stations},
which also contains additional information on how these codes were developed.
}
\usage{
bw_parameters(continuity = NULL)
}
\arguments{
\item{pars}{Optional: if empty all available parameters will be returned.
Alternatively, a vector of the continuous or discrete parameters can be
requested.}
}
\value{
A vector of parameters.
}
\description{
\code{bw_parameters} returns a vector of parameters that can be retrieved from
Water Data Online.
}
\details{
The units of the parameters are as follows:
\itemize{
\item Water Course Discharge (m3/s)
\item Water Course Level (m)
\item Electrical conductivity at 25C (µS/cm)
\item Turbidity (NTU)
\item pH
\item Water Temperature (ºC)
\item Storage Volume (ML)
\item Storage Level (m)
\item Ground Water Level (m)
\item Rainfall (mm)
\item Evaporation (mm)
\item Dry Air Temperature (ºC)
\item Relative Humidity (\%)
\item Wind Speed (m/s)
}
}
\examples{
bw_parameters()
bw_parameters("continuous")
bw_parameters("discrete")
}
\seealso{
\itemize{
\item \url{http://www.bom.gov.au/waterdata/}
\item \href{http://www.bom.gov.au/waterdata/wiski-web-public/Guide\%20to\%20Sensor\%20Observation\%20Services\%20(SOS2)\%20for\%20Water\%20Data\%20\%20Online\%20v1.0.1.pdf}{BoM Guide to Sensor Observation Services (SOS2) for Water Data Online}
}
}
