% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{bw_quality_codes}
\alias{bw_quality_codes}
\title{Access the bomWater quality code conversion table}
\usage{
bw_quality_codes()
}
\value{
A tibble with three columns: Quality Code, BOM Quality Code, and
Description.
}
\description{
\code{quality_codes()} returns a table which can be used to convert the numeric
data quality codes returned by the WISKI API into the human readable codes
used on Water Data Online.
}
\examples{
quality_codes()
}
