% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{bw_return_fields}
\alias{bw_return_fields}
\title{\itemize{
\item station_no',
\item station_id
\item station_name
\item stationparameter_id
\item stationparameter_no
\item stationparameter_name
\item stationparameter_longname
\item site_id
\item site_no
\item site_name
\item parametertype_id
\item parametertype_name
\item parametertype_longname
\item parametertype_unitname
\item parametertype_shortunitname
}}
\usage{
bw_return_fields(request)
}
\description{
\itemize{
\item station_no',
\item station_id
\item station_name
\item stationparameter_id
\item stationparameter_no
\item stationparameter_name
\item stationparameter_longname
\item site_id
\item site_no
\item site_name
\item parametertype_id
\item parametertype_name
\item parametertype_longname
\item parametertype_unitname
\item parametertype_shortunitname
}
}
