% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list.R
\name{get_parameter_list}
\alias{get_parameter_list}
\title{Retrieve available parameters for stations}
\usage{
get_parameter_list(
  station_number,
  return_fields = c("station_no", "station_name", "parametertype_name",
    "parametertype_unitname")
)
}
\arguments{
\item{station_number}{A single or multiple vector of AWRC station numbers.}

\item{return_fields}{(Optional) Station parameter details to be returned. By
default the return fields are: station_no, station_id, station_name,
parametertype_id, parametertype_name, parametertype_unitname
parametertype_shortunitname.}
}
\value{
A tibble with columns for each of the return fields.
}
\description{
\code{get_parameter_list} returns the parameters that can be
retrieved at one or more stations.
}
\details{
The default return fields have been selected to generally return the
most useful fields while reducing duplication of metadata. The full list of
return fields:
}
\examples{
# Return parameters for a single station
\dontrun{
get_parameter_list(station_number = "410730")
}
# Return available parameters for multiple stations
\dontrun{
get_parameter_list(station_number = c("410730", "570946"))
}
}
