% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list.R
\name{get_station_list}
\alias{get_station_list}
\title{List water observation stations}
\usage{
get_station_list(
  parameter = "Water Course Level",
  station_number = NULL,
  bbox = NULL,
  return_fields = c("station_no", "station_name", "station_latitude",
    "station_longitude")
)
}
\arguments{
\item{parameter}{The parameter for the station request. Defaults to
"Water Course Level".}

\item{station_number}{Optional: a vector of AWRC station numbers.}

\item{bbox}{Optional: a bounding box to get stations in a region of interest.
Takes a vector ordered xmin, ymin, xmax, ymax.}

\item{return_fields}{A character vector of station details to be returned.
See [return_fields()] for a full list of available fields.}
}
\value{
A tibble with columns gven by return_fields.
}
\description{
`get_station_list` queries Water Data Online and returns station details.
Queries can be input with the desired `parameter` to find all the
stations on record. If you already have a vector of station numbers, you can
pass the vector to `station_number` and return the details of those stations.
`return_fields` can be customised to return various data about the stations.
}
\examples{
# Get a list of all Water Course Discharge Stations
\dontrun{
get_station_list(parameter = "Water Course Discharge")
}
# Just get the details for Cotter River at Gingera
\dontrun{
get_station_list(station_number = "410730")
}
# Get all attributes for one station
\dontrun{
all_discharge_fields <- return_fields("Water Course Discharge")
get_station_list("Water Course Discharge", "410730", return_fields = all_discharge_fields)
}
}
