% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomWater.R
\name{get_timeseries_id}
\alias{get_timeseries_id}
\title{Retrieve the timeseries ID}
\usage{
get_timeseries_id(parameter_type, station_number, ts_name)
}
\arguments{
\item{parameter_type}{The parameter of interest (e.g. Water Course
Discharge).}

\item{station_number}{The AWRC station number.}

\item{ts_name}{The BoM time series name (e.g. DMQaQc.Merged.DailyMean.24HR).}
}
\value{
Returns a tibble with columns station_name, station_no, station_id, ts_id,
ts_name, parametertype_id, parametertype_name.
}
\description{
`get_timeseries_id` retrieves the timeseries ID that can be used to obtain
values for a parameter type, station and timeseries combination.
}
\author{
Alexander Buzacott
}
