% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomWater.R
\name{get_timeseries_values}
\alias{get_timeseries_values}
\title{Retrieve timeseries values}
\usage{
get_timeseries_values(ts_id, start_date, end_date, return_fields)
}
\arguments{
\item{ts_id}{The timeseries ID for the values of interest. Can be found using
the function `get_timeseries_id`.}

\item{start_date}{The start date formatted as 'YYYY-MM-DD'.}

\item{end_date}{The end date formatted as 'YYYY-MM-DD'.}

\item{return_fields}{A vector of the variables that are to be returned.}
}
\value{
A tibble with columns with the requested return_fields. A zero row tibble is
returned if no data is returned  from the query. The columns of the tibble
are returned as character classes and have not been formatted to more
appropriate correct classes (this happens in other functions).
}
\description{
`get_timeseries_values` returns the timeseries values between a start and end
date for given timeseries ID.
}
\author{
Alexander Buzacott
}
