% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomWater.R
\name{getStationList}
\alias{getStationList}
\title{Retrieve water observation stations}
\usage{
getStationList(parameterType, stationNumber, returnFields)
}
\arguments{
\item{parameterType}{The parameter for the station request (e.g. Water Course Discharge,
Storage Level)}

\item{stationNumber}{Optional: a single or multiple vector of AWRC station numbers.}

\item{returnFields}{Station details to be returned. By default the columns returned are
station name, number, ID, latitude and longitude.
 Can be customised with a vector of parameters.}
}
\value{
With the default return fields, a tibble with columns station_name, station_no, station_id,
station_latitude, station_longitude.
}
\description{
`getStationList` queries Water Data Online and returns station details. Queries can be
input with the desired `parameterType` to find all the stations on record. If you already
have a vector of station numbers, you can pass the vector to `stationNumber` and return the
details of those stations. `returnFields` can be customised to return various data about
the stations.
}
\examples{
# Get all Water Course Discharge Stations
getStationList()
# Just the details for Cotter River at Gingera
getStationList(stationNumber='410730')
# Rainfall stations
getStationList(parameterType = 'Rainfall')

}
\author{
Alexander Buzacott
}
