% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeries.R
\name{getTimeseries}
\alias{getTimeseries}
\title{Get formatted timeseries data}
\usage{
getTimeseries(
  parameterType,
  stationNumber,
  startDate,
  endDate,
  tz,
  returnFields,
  tsName
)
}
\arguments{
\item{parameterType}{The water data parameter type (e.g. Water Course Discharge). See below for
a list of available parameters}

\item{stationNumber}{The AWRC station number.}

\item{startDate}{Start date formatted as a string or date class (YYYY-MM-DD)}

\item{endDate}{End date formatted as a string or date class (YYYY-MM-DD)}

\item{tz}{Optional: the desired time zone for the output timeseries. Input must be an Olson Name
(see \code{OlsonNames()}). By default the the timeseries is returned in an offset timezone
(e.g. \code{Etc/GMT-10} for NSW) as the timeseries do not observe DST.}

\item{returnFields}{Optional: columns to be returned from Water Data Online. By default Timestamp
Value and Quality Code are returned.}

\item{tsName}{The timeseries name (e.g. DMQaQc.Merged.DailyMean.24HR) that is desired.}
}
\value{
A tibble with the requested return fields, which by default are \code{Timestamp}, \code{Value} and \verb{Quality Code}.
Zero row tibbles are returned if no data is available for the requested dates. The aggregation of data
is generally the mean for most variables, except for rainfall and evaporation which is the sum over the chosen period.
}
\description{
This function returns a formatted timeseries from Water Data Online.
}
\details{
This function returns a timeseries within the chosen date window. The units of the timeseries values are as per
Water Data Online and can be viewed on page 39 of the Water Data Online SOS manual (linked in See Also below).

The following parameters (\code{parameterType}) can be requested:
\itemize{
\item Water Course Discharge
\item Water Course Level
\item Storage Volume
\item Storage Level
\item Ground Water Level
\item Rainfall
\item Evaporation
\item Dry Air Temperature
\item Relative Humidity
\item Wind Speed
\item Electrical conductivity at 25C
\item Turbidity
\item Water pH
\item Water Temperature
}

This function can be used if the others have missed a particular time series (e.g. the forecasting timeseries)
}
\examples{
\dontrun{
getHourlyData(parameterType = 'Water Course Discharge',
              stationNumber = '410730',
              startDate     = '2020-01-01',
              endDate       = '2020-01-31')
}
\dontrun{
getDailyData(parameterType = 'Water Course Discharge',
             stationNumber = '410730',
             startDate     = '2020-01-01',
             endDate       = '2020-01-31',
             var           = 'Mean',
             aggregation   = '09HR')
}
\dontrun{
getYearly(parameterType = 'Rainfall',
          stationNumber = '570946',
          startYear     = 2016,
          endYear       = 2020)
}
}
\seealso{
\url{http://www.bom.gov.au/waterdata/}
\url{http://www.bom.gov.au/waterdata/wiski-web-public/Guide\%20to\%20Sensor\%20Observation\%20Services\%20(SOS2)\%20for\%20Water\%20Data\%20\%20Online\%20v1.0.1.pdf}
}
\author{
Alexander Buzacott
}
