% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomWater.R
\name{getTimeseriesID}
\alias{getTimeseriesID}
\title{Retrieve the timeseries ID}
\usage{
getTimeseriesID(parameterType, stationNumber, tsName)
}
\arguments{
\item{parameterType}{The parameter of interest (e.g. Water Course Discharge).}

\item{stationNumber}{The AWRC station number.}

\item{tsName}{The BoM time series name (e.g. DMQaQc.Merged.DailyMean.24HR).}
}
\value{
Returns a tibble with columns station_name, station_no, station_id, ts_id, ts_name,
parametertype_id, parametertype_name.
}
\description{
`getTimeseriesID` finds the timeseries ID that can be used to obtain values for a parameter type,
station and timeseries combination.
}
\examples{
\dontrun{
getTimeseriesID(parameterType = 'Water Course Discharge',
                stationNumber = '410730',
                tsName = 'DMQaQc.Merged.DailyMean.24HR')
}
}
\author{
Alexander Buzacott
}
