% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomWater.R
\name{getTimeseriesValues}
\alias{getTimeseriesValues}
\title{Retrieve timeseries values}
\usage{
getTimeseriesValues(tsID, startDate, endDate, returnFields)
}
\arguments{
\item{tsID}{The timeseries ID for the values of interest. Can be found using the function
`getTimeseriesID`.}

\item{startDate}{The start date formatted as 'YYYY-MM-DD'.}

\item{endDate}{The end date formatted as 'YYYY-MM-DD'.}

\item{returnFields}{A vector of the variables that are to be returned.}
}
\value{
A tibble with columns with the requested returnFields. A zero row tibble is returned if no data
is returned  from the query. The columns of the tibble are returned as character classes and
have not been formatted to more appropriate correct classes (this happens in other functions).
}
\description{
`getTimeseriesValues` returns the timeseries values between a start and end date for given
timeseries ID.
}
\examples{
\dontrun{
# Get the timeseries ID
timeseriesID = getTimeseriesID(parameterType = 'Water Course Discharge',
                               stationNumber = '410730',
                               tsName = 'DMQaQc.Merged.DailyMean.24HR')

getTimeseriesValues(timeseriesID$ts_id,
                    startDate,
                    endDate,
                    returnFields = c('Timestamp',
                                     'Value',
                                     'Quality Code',
                                     'Interpolation Type'))
}

}
