% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTables.R
\name{interpolation_codes}
\alias{interpolation_codes}
\title{Access the interplation codes conversion table}
\usage{
interpolation_codes()
}
\value{
a tibble with three columns: Interpolation Type, Name, and
Description
}
\description{
`interpolation_codes()` returns a table which can be used to interpret the
interpolation codes returned by the WISKI API.
}
\examples{
interpolation_codes()
}
