% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTables.R
\name{interpolation_types}
\alias{interpolation_types}
\title{Access the interplation types conversion table}
\source{
Interpolation types are detailed in the
\href{http://www.bom.gov.au/waterdata/wiski-web-public/Guide\%20to\%20Sensor\%20Observation\%20Services\%20(SOS2)\%20for\%20Water\%20Data\%20\%20Online\%20v1.0.1.pdf}{BOM SOS2 manual}
(6.3.2, pg 40).
}
\usage{
interpolation_types()
}
\value{
a tibble with three columns: Interpolation Type, Name, and
Description.
}
\description{
`interpolation_types()` returns a table which can be used to interpret the
interpolation types returned by the WISKI API.
}
\examples{
interpolation_types()
}
