% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{make_bom_request}
\alias{make_bom_request}
\title{Query the BOM WISKI API}
\usage{
make_bom_request(params, tries = 1)
}
\arguments{
\item{params}{A named list of parameters.}

\item{tries}{Number of times to query the API. Only a single call is made by
default, and the maximum number of tries is 5.}
}
\value{
A tibble with columns depending on the request. For
  \code{\link{get_timeseries}} requests, a tibble with zero rows is returned
  if there is no data available for that query.
}
\description{
This function queries the Bureau of Meteorology (BOM) Water Data KISTERS API.
A parameter list is passed to make a the request and the JSON return is parsed
depending on what is requested. This function can be used if you want to
build your own JSON queries.
}
