% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomWater.R
\name{makeBOMRequest}
\alias{makeBOMRequest}
\title{Query the BoM WISKI API}
\usage{
makeBOMRequest(params)
}
\arguments{
\item{params}{A named list of parameters.}
}
\value{
A tibble is returned with the columns depending on the request. For \code{getTimeSeries}
requests, a tibble with zero rows is returned if the request if there is no data
available for that query.
}
\description{
This function queries the Bureau of Meteorology Water Data KISTERS API.
A parameter list is passed to make request and the JSON return is parsed depending
on what is requested. This function can be used if you want to build your own JSON queries.
}
\examples{
# Getting the stations for Water Course Discharge
\dontrun{
params = list('request' = 'getStationList',
              'parameterType_name' = 'Water Course Discharge',
              'returnfields' = paste(c('station_name',
                                       'station_no',
                                       'station_id',
                                       'station_latitude',
                                       'station_longitude'),
                                     collapse=','))
makeBOMRequest(params)
}
}
\author{
Alexander Buzacott
}
