% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTables.R
\name{quality_codes}
\alias{quality_codes}
\title{Access the quality code conversion table}
\usage{
quality_codes()
}
\value{
a tibble with three columns: Quality Code, BOM Quality Code, and Description
}
\description{
`quality_codes()` returns a table which can be used to convert the numeric
data quality codes returned by the WISKI API into the human readable codes
used on Water Data Online.
}
\examples{
quality_codes()
}
