% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTables.R
\name{interpolation_types}
\alias{interpolation_types}
\title{Access the interplation types conversion table}
\usage{
interpolation_types()
}
\value{
a tibble with three columns: Interpolation Type, Name, and
Description
}
\description{
`interpolation_types()` returns a table which can be used to interpret the
interpolation types returned by the WISKI API.
}
\examples{
interpolation_types()
}
