% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomWater.R
\name{make_bom_request}
\alias{make_bom_request}
\title{Query the BoM WISKI API}
\usage{
make_bom_request(params)
}
\arguments{
\item{params}{A named list of parameters.}
}
\value{
A tibble is returned with the columns depending on the request. For
\code{get_timeseries} requests, a tibble with zero rows is returned
if there is no data available for that query.
}
\description{
This function queries the Bureau of Meteorology Water Data KISTERS API.
A parameter list is passed to make request and the JSON return is parsed
depending on what is requested. This function can be used if you want to
build your own JSON queries.
}
\author{
Alexander Buzacott
}
