% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_request.R
\name{make_bom_request}
\alias{make_bom_request}
\title{Query the BOM WISKI API}
\usage{
make_bom_request(params, max_tries = 1)
}
\arguments{
\item{params}{A named list of key-value pairs.}

\item{max_tries}{Optional number number of times to repeat the request if an
error is returned. Silently capped at 5 tries.}
}
\value{
A tibble with columns depending on the request. For
\code{\link{get_timeseries}} requests, a tibble with zero rows is returned
if the query has no result.
}
\description{
\code{make_bom_request} queries the Bureau of Meteorology (BOM) Water Data KISTERS
API and returns the result as a tibble. This function has useful wrappers for
common requests (refer to See Also below), however it can be used directly if
you want to build a custom query.
}
