% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTables.R
\name{quality_codes}
\alias{quality_codes}
\title{Access the quality code conversion table}
\source{
The conversion of SOS2 quality codes into BOM quality codes is from the
\href{http://www.bom.gov.au/waterdata/wiski-web-public/Guide\%20to\%20Sensor\%20Observation\%20Services\%20(SOS2)\%20for\%20Water\%20Data\%20\%20Online\%20v1.0.1.pdf}{BOM SOS2 manual}
(6.3.3, pg 41). Quality code descriptions are from
\href{http://www.bom.gov.au/water/hrs/qc_doc.shtml}{Streamflow data quality codes for Hydrologic Reference Stations},
which also contains additional information on how these codes were developed.
}
\usage{
quality_codes()
}
\value{
A tibble with three columns: Quality Code, BOM Quality Code, and Description.
}
\description{
`quality_codes()` returns a table which can be used to convert the numeric
data quality codes returned by the WISKI API into the human readable codes
used on Water Data Online.
}
\examples{
quality_codes()
}
